Modifying how connection closing is handled
Several changes:
1) Moving all the connection-closing-related logic to the
._handle_closing() method: This encapsulate all this logic in a single
method instead of scattering it on every method calling
._handle_closing()
2) Explicitly closing the cursor: on the previous version,
we didn't cared about closing the cursor explicitly because it happened
when the connection was closed. However, as now we need to care about
cases where the cursor may be closed independently from the connection,
we need to make it explicit
3) Adding a new argument to ._handle_closing(): in order to distinguish
between the connection closings and the connection terminations, we have
added a `force` argument to the method. It is False by default, but it
should be set to True when we want to be sure that the connection is
getting closed no matter what (e.g. in the terminate method).